js/site.js: Fix broken "Copy link" button on share dialog
The button only breaks locally as insecure contexts don't have access
to the Clipboard API (see
https://developer.mozilla.org/en-US/docs/Web/API/Clipboard_API for more
info), and so does a fallback to selecting the text using the given HTML
element. However, as the HTML element passed is actually the input
element's **value**, it'll hence throw an error saying that "select() is
not defined" (as `select()` only works on input elements, not strings).